Split out page object classes (DH-18428-2)
