+++ e2e-testing/src/pageObjects/index.ts
@@ -1,0 +1,2 @@
+export * from './EditorViewExtended';
+export * from './WebViewExtended';
