+++ e2e-testing/src/panels.spec.ts
@@ -1,16 +1,15 @@
-import { EditorView, WebView, Workbench } from 'vscode-extension-tester';
+import { EditorView, Workbench } from 'vscode-extension-tester';
 import path from 'node:path';
 import {
-  EditorViewExtended,
   elementExists,
   elementIsLazyLoaded,
-  extendWebView,
   getCodeLens,
   openFileResources,
   type EditorGroupData,
 } from './testUtils';
 import { assert } from 'chai';
 import { locators } from './locators';
+import { EditorViewExtended } from './pageObjects';
 
 const testWsPath = path.resolve(__dirname, '..', 'test-ws/');
 const simpleTicking3Name = 'simple_ticking3.py';
@@ -83,28 +82,25 @@ describe('Panels Tests', () => {
     assert.deepEqual(editorGroupsData, expectedGroupState.initial);
 
     // Tab 3 should be selected already since it was the last query to be run.
-    // Get the WebView without selecting the tab since we expect it to already
-    // be selected
-    const group2 = await editorView.getEditorGroup(2);
-    let t3 = extendWebView(await new WebView(group2).wait());
+    let t3 = await editorView.openWebView(2, 't3', true);
     await t3.switchToContentFrame();
     assert.isTrue(
       await elementIsLazyLoaded(locators.irisGrid),
-      'Iris grid should be lazy loaded after panel becomes active'
+      'Iris grid should be automatically loaded for initially active panel'
     );
     await t3.switchBack();
 
     // Switch to tab 1
-    const t1 = await editorView.openWebView('t1', 2);
+    const t1 = await editorView.openWebView(2, 't1');
     await t1.switchToContentFrame();
     assert.isTrue(
       await elementIsLazyLoaded(locators.irisGrid),
-      'Iris grid should be lazy loaded after panel becomes active'
+      'Iris grid should be loaded after panel becomes active'
     );
     await t1.switchBack();
 
     // Switch back to tab 3
-    t3 = await editorView.openWebView('t3', 2);
+    t3 = await editorView.openWebView(2, 't3');
     await t3.switchToContentFrame();
     assert.isTrue(
       await elementExists(locators.irisGrid),
